requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
boto3
python-dotenv
langchain==0.1.14
langchain-community>=0.0.31
langchain-core==0.1.50
coloredlogs>=15.0.1
jq==1.7.0
opensearch-py==2.5.0
loguru==0.7.2
streamlit
